### 2025-08-13 #8794 @danielskatz
Also, given the short commit history in the repo, can you discuss the amount of work you did before the first commit?
