### 2025-08-22 #8794 @cfosseprez
@danielskatz sorry for the late reply I was away.

I have updated the license using the [link you provided](https://choosealicense.com/licenses/gpl-3.0/).

Regarding the short commit history, before the first commit of this repository, I developed the core of the code for about a month, and used it for a year or so. Recently I did the factoring and integrated more features that were previously in separate chuck of code. So in total its about 2-3 month of work for the coding, plus several other months of experimenting to define the constraints and identify the need.

Best
