+++ paper/paper.bib
@@ -3,14 +3,16 @@
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
   journal={Proceedings of the 7th Python in Science Conference},
   pages={11--15},
-  year={2008}
+  year={2008},
+  doi = {10.25080/TCWV9851}
 }
 
 @book{krause2002living,
   title={Living in groups},
   author={Krause, Jens and Ruxton, Graeme D},
   year={2002},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/oso/9780198508175.001.0001}
 }
 
 @article{csardi2006igraph,
@@ -18,5 +20,6 @@
   author={Csardi, Gabor and Nepusz, Tamas},
   journal={InterJournal Complex Systems},
   volume={1695},
-  year={2006}
+  year={2006},
+  doi = {10.5281/zenodo.3630268}
 }
