### 2025-09-03 #8794 @vissarion
Hi @diningphil @szhorvat @paulbrodersen @IvanIsCoding would any of you be willing to review this submission for JOSS?

As you probably know we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
