### 2025-09-03 #8794 @szhorvat
@vissarion  I recommend @Zepeacedust as a reviewer, which would help the review be completed faster than then 4-6 weeks I mentioned. He is a student working with me. He implemented the soon-to-be-released spatial network features in igraph, and he has the necessary expertise.
