### 2025-09-08 #8794 @vissarion
Hi @Zepeacedust would you be willing to review this submission for JOSS?

As @szhorvat recommended, you have the right background for this. In JOSS we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
