### 2025-10-26 #9304 @editorialbot
[REVIEW]: Graphizy: Memory-Enhanced Graph Construction for Computational Geometry and Temporal Network Analysis
**Submitting author:** @cfosseprez (<a href="http://orcid.org/0009-0000-4524-0399">Charles Fosseprez</a>)
**Repository:** https://github.com/cfosseprez/graphizy
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.20
