### 2025-11-14 #9304 @IvanIsCoding
I will still complete the review. But compared to other JOSS reviews, the pre-review for this had an unfavourable 6 week gap that covered when I had budgeted for the review.

I will try to get initial comments within a week and a complete review by December 1st.
