### 2026-01-22 #9304 @IvanIsCoding
This has been a very abnormal review. I have not completed it, but it seems the other reviewer has found issues that have not been closed yet? https://github.com/openjournals/joss-reviews/issues/9304#issuecomment-3665200658
