### 2026-01-22 #9304 @vissarion
Hi, let us get some status update:

  @Zepeacedust has raised a scholarly effort issue and created some issues mostly on documentation on the repository under submission. @Zepeacedust does the reply by the authors resolve your scholarly effort issues?

  @cfosseprez could you please address the issues created in the repository, so that the review can move forward?

  @IvanIsCoding you have partially review the submission, are there any issues that hold you back?
