### 2026-02-17 #9304 @vissarion
Hi all, it seems that this submission is stuck. To my understanding the major issue is the scholarly effort raised by @Zepeacedust. There is a reply by the authors https://github.com/openjournals/joss-reviews/issues/9304#issuecomment-3679960138 on that. @Zepeacedust what is your opinion on this?
