### 2026-02-17 #9304 @Zepeacedust
I am not certain, but I would err on the side of no.
Relatively little original work has been done here, with most of the effort being a united interface.
As I understand the guidelines, that does not constitute scholarly effort.
