### 2026-03-03 #9304 @Zepeacedust
> @Zepeacedust I understand you concerns, but according to previous scope JOSS guidelines a wrapper bringing multiple packages together under a unified interface with enough LOC (here we have >10K) is acceptable. If this OK with you could you please continue your review according to that guidelines?
>

In that case, I would say that it is acceptable, but I eould want better documentation in the future.
