### 2026-03-04 #9304 @vissarion
Hi, @edeno @amadeusine @KennethKwabenaKenney @Jacob-Barhak @amritagos we need one more reviewer here. Would any of you be willing to review this submission for JOSS?

As you probably know we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
