### 2026-03-04 #9304 @Jacob-Barhak
Sorry Vissarion,

Unless JOSS changes its policies to recognize public domain licenses such
as CC0 then I cannot support your cause.
I do support public non blind review, yet if your definition of open is
biased, I cannot support you.
Look at this discussion:
https://github.com/openjournals/joss/issues/889

To give you some food for thought, here is an example of someone who
decided to waive their copyright and did a lot of good.
https://luc.devroye.org/rnbookindex.html

The world is changing. Time to catch up.

Please reach out when you reconsider your definition of open source.

                 Jacob

On Wed, Mar 4, 2026 at 2:36 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> *vissarion* left a comment (openjournals/joss-reviews#9304)
> <https://github.com/openjournals/joss-reviews/issues/9304#issuecomment-3996062105>
>
> Hi, @edeno <https://github.com/edeno> @amadeusine
> <https://github.com/amadeusine> @KennethKwabenaKenney
> <https://github.com/KennethKwabenaKenney> @Jacob-Barhak
> <https://github.com/Jacob-Barhak> @amritagos
> <https://github.com/amritagos> we need one more reviewer here. Would any
> of you be willing to review this submission for JOSS?
>
> As you probably know we carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so
> kind to let me know even if you're not interested, so that I can start
> looking for other reviewers.
>
