### 2026-03-04 #9304 @vissarion
Jacob, thanks for your reply and the links. This is definitely an interesting and important discussion.
Let's focus on topic of the review here and continue this discussion in other channels.
