### 2026-03-04 #9304 @Jacob-Barhak
Sure Vissarion,

This discussion was ongoing for a long time. You can see details in slide
35:
https://www.clinicalunitmapping.com/show/AI_legal_Latest.pdf

This discussion ended here:
https://groups.google.com/a/numfocus.org/g/licensing

If you truly want to discuss this, perhaps it is a good place to continue
discussion.

            Jacob

On Wed, Mar 4, 2026 at 3:12 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> *vissarion* left a comment (openjournals/joss-reviews#9304)
> <https://github.com/openjournals/joss-reviews/issues/9304#issuecomment-3996229431>
>
> Jacob, thanks for your reply and the links. This is definitely an
> interesting and important discussion.
> Let's focus on topic of the review here and continue this discussion in
> other channels.
>
