### 2026-03-27 #9304 @KennethKwabenaKenney
Hello editor @vissarion this is the summary from my review

**SUMMARY**
---
This paper introduces Graphizy, a Python package designed for constructing and analyzing spatial-temporal graphs with a temporal memory system. The software addresses key limitations in existing tools such as NetworkX and igraph, particularly their lack of native support for temporal persistence in evolving networks.
The contribution is relevant and potentially impactful, especially for domains involving dynamic spatial interactions

Also it seem to be that the submitting Author is the only Contributor

**MAJOR STRENGTHS**
1. Clear Statement of Need
The paper clearly identifies three gaps in current tools: Lack of spatial-first graph construction, no temporal persistence tracking, Fragmented multi-library workflows. These are well-articulated and justified.
2. Novel Temporal Memory Concept
The temporal memory system is the core innovation.
Which enables:
Tracking edge persistence
Distinguishing transient vs stable relationships
This is scientifically meaningful, especially for noisy real-world data.

**MAJOR & MINOR CONCERNS**
1. Performance claims (e.g., 18× average speedup) lack benchmark setup details such as hardware specifications and dataset descriptions

2. Insufficient Validation Detail: the paper states that “validated against established implementations” but no quantitative comparison metrics, error bounds or similarity metrics are shown

3. The “temporal memory system” is central, so i suggest there should be a formal definition or equations or algorithmic description

4.  the statement “impossible with traditional approaches” is too strong, i suggest rephrasing it to “not natively supported”
